Adding g++ and newman installation to playbook
If Newmen proves useful these tasks will form the start of a Newman role.

Currently we can't rely on the Ansible NPM module as we need to use the 'unsafe-perm' parameter. A pull request for adding this to the module is currnetly open (not by us) and pending merging so this should be temporary.