Adding base JSON schema to Postman tests
This updated collection includes a check against the basic JSON schema created for the POST tokens method.

Annoying the schema has to be hard-coded within the collection due to a limitation with Postman being a Chrome application. This is being worked on but until then we will have to duplicate schemas into collections.

Exporting the collection and running through Newman proved successful, though how errors are handled (i.e. displayed) is not clear when run from the CLI. In the main application you need to open dev-tools to see the console, which obviously don't exist in the CLI.

For now re-running a failed test in Postman to debug errors will be acceptable. If Postman has not provided a solution in the future we could work on ways to do the schema validation outside of Newman as a separate test step.