Adding descriptions and examples
'Example' is not defined within the JSON schema spec, but is not excluded either. 'Default' is defined in the spec but doesn't make sense for the properties it's been added to. Notices won't always have a type of 'token_generated' and it wouldn't make sense to assume this for example.

We can use this keyword in API documentation for example, alongside Default if appropriate.